requirements.txt
# requirements.txt
"""
# Core ML/NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
spacy>=3.6.0
scispacy>=0.5.0

# Langchain ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0

# Medical NLP specific
datasets>=2.12.0
tokenizers>=0.13.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
unstructured>=0.10.0

# Web interface (optional)
streamlit>=1.28.0
gradio>=3.40.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
"""
